started working on entity module

+++ LotusClient/Core_Engine/BaseClasses/Types/Velocity.cs
@@ -4,22 +4,22 @@ public class Velocity
 {
     public float _PerTickMulitplier = (1 / 8000);
 
-    public short _XVel;
-    public short _YVel;
-    public short _ZVel;
+    public double _XVel;
+    public double _YVel;
+    public double _ZVel;
 
     public float GetXMoveForSingleTick()
     {
-        return _PerTickMulitplier * _XVel;
+        return (float)(_PerTickMulitplier * _XVel);
     }
 
     public float GetYMoveForSingleTick()
     {
-        return _PerTickMulitplier * _YVel;
+        return (float)(_PerTickMulitplier * _YVel);
     }
 
     public float GetZMoveForSingleTick()
     {
-        return _PerTickMulitplier * _ZVel;
+        return (float)(_PerTickMulitplier * _ZVel);
     }
 }
